+++ src/enclave/client_session.h
@@ -7,7 +7,7 @@
 
 namespace ccf
 {
-  class ClientEndpoint
+  class ClientSession
   {
   public:
     using HandleDataCallback = std::function<void(
@@ -23,26 +23,27 @@ namespace ccf
     HandleErrorCallback handle_error_cb;
 
   private:
-    int64_t session_id;
+    int64_t client_session_id;
     ringbuffer::WriterPtr to_host;
 
   public:
-    ClientEndpoint(
-      int64_t session_id, ringbuffer::AbstractWriterFactory& writer_factory) :
-      session_id(session_id),
+    ClientSession(
+      int64_t client_session_id,
+      ringbuffer::AbstractWriterFactory& writer_factory) :
+      client_session_id(client_session_id),
       to_host(writer_factory.create_writer_to_outside())
     {}
 
-    virtual void send_request(const http::Request& request) = 0;
+    virtual void send_request(http::Request&& request) = 0;
 
-    void connect(
+    virtual void connect(
       const std::string& hostname,
       const std::string& service,
       const HandleDataCallback f,
       const HandleErrorCallback e = nullptr)
     {
       RINGBUFFER_WRITE_MESSAGE(
-        tls::tls_connect, to_host, session_id, hostname, service);
+        tls::tls_connect, to_host, client_session_id, hostname, service);
       handle_data_cb = f;
       handle_error_cb = e;
     }